move the class to collisions.py file
